Split app.js into entry wiring + focused UI modules (1239 -> 343 lines)
- Extract notice.js, side-panels.js, user-def-modal.js, layer-modal.js
- Render user-def tables from shared column definitions; replace per-row
  listeners with single event delegation
- Use dom-utils markInputInvalid/clearInputInvalid
- Invalidate cssVar cache on theme change
